backend/requirements.txt
fastapi
uvicorn
mysql-connector-python
python-multipart
bcrypt
PyJWT
pdfplumber
sentence-transformers
faiss-cpu
numpy
python-dotenv